requirements-llm.txt
arabic-reshaper==3.0.0
beautifulsoup4==4.12.3
certifi==2024.8.30
chardet==3.0.4
charset-normalizer==3.4.0
deep-translator==1.11.4
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.11.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
pysrt==1.1.2
python-bidi==0.6.3
requests==2.32.3
rfc3986==1.5.0
setuptools==75.1.0
sniffio==1.3.1
soupsieve==2.6
tqdm==4.66.6
urllib3==2.2.3
wheel==0.44.0
torch
torchaudio
transformers
sentencepiece
protobuf
